reorganize files to hopefully be more sane
