--- db/index.js
@@ -1,4 +1,0 @@
-var P = require('p-promise')
-var kvstore = require('kvstore')
-
-module.exports = require('./kv')(P, kvstore)
