--- log/index.js
@@ -1,3 +1,0 @@
-var bunyan = require('bunyan')
-
-module.exports = require('./log')(bunyan)
