+++ memory_monitor.js
