+++ test/account_tests.js
@@ -1,17 +1,16 @@
 var test = require('tap').test
+var kvstore = require('kvstore')
 var P = require('p-promise')
 
 function FakeToken() {}
-
 FakeToken.get = function () { return P(new FakeToken())}
-
 FakeToken.prototype.del = function () { return P(null) }
 
 function FakeRecoveryMethod() {}
 FakeRecoveryMethod.create = function () { return P(new FakeRecoveryMethod()) }
 FakeRecoveryMethod.get = function () { return P(new FakeRecoveryMethod()) }
 
-var db = require('../db')({
+var db = require('../models/kv')(P, kvstore, {
   kvstore: {
     available_backends: ['memory'],
     backend: 'memory',
@@ -19,7 +18,7 @@ var db = require('../db')({
   }
 })
 var DOMAIN = 'example.com'
-var Account = require('../account')(P, FakeToken, FakeRecoveryMethod, db.store, DOMAIN)
+var Account = require('../models/account')(P, FakeToken, FakeRecoveryMethod, db.store, DOMAIN)
 var a = {
   uid: 'xxx',
   email: 'somebody@example.com',
